extract downloads.markDownloadMissingIfGone helper
